Refactoring incomplet
